// src/models/addClient.ts
export interface IaddClient {
    address: string
    city: string
    company_name: string
    company_type: string
    email: string
    industry_id: number
    name: string
    primary_number: string
    secondary_number: string
    user_id: number
    vat: string
    zipcode: string
}